Updated Marker Location
Displaying the text by the REDCap logo was nice.  However the logo and
top navigation bar change size.  The Instance Marker was moved to the
top right corner instead of the top left.

Better directions, thanks Rob.

Basic debug info can be turned on and display in the console log.

Improvement: stripping tags from the user speicific text.  The idea is
to keep this as small, yet as visually obvious as possible.

Older code was commented out.

CSS is appended to the head instead of using inline css.